Remove NPG tables from project
The NPG tables aren't directly related to the warehouse builders.
Keeping them as part of this project will mean that schema changes
to the table will need to be made through the builders, or that the
structure and actual schema of the tables will get out of sync.

In addition, this allows NPG to populate their own tables in a
given environment before the builders are deployed, without risking
us dropping the tables.