Switch contains/contaminated_human_dna to bools
The columns contains_human_dna and
contaminated_human_dna were effectively booleans,
yet Sequencescape was broadcasting 'Yes' and 'No'.
This was inconsistent with other columns in the
same table, but updating Sequencescape will have
knock on effects.

- Two migrations to convert data, and update the column

- Add store_as_boolean for string conversion

store_as_boolean ensures that 'Yes' and 'No' are
converted to true and false booleans respectively.