Do not record tag2 index
Downstream processes require a single identifier to reference an aliquot
within a pool. Further constraints require that this be the tag_index field.
Given that the behaviour of this field has now changed, and now merely reflects
the index of an aliquot within the pool, tag2 index would be even more
misleading.